+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Install the `pgvector` extension in the local Postgres container.
+
 ## [0.1.0] - 2024-03-21
 
 ### Added